backend/.history/requirements_20260117182929.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
pydantic>=2.7.4
pydantic-settings>=2.4.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymysql==1.1.0
mysql-connector-python==9.5.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
openai==1.72.0
dashscope==1.25.5
python-dotenv==1.0.1
email-validator==2.3.0
dnspython==2.8.0
faiss-cpu==1.13.2
tavily-pytho